% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GStats.R
\docType{methods}
\name{aggregate,GStats-method}
\alias{aggregate,GStats-method}
\title{Aggregate GStats statistics}
\usage{
\S4method{aggregate}{GStats}(x, formula, FUN, per.site = FALSE, ...)
}
\arguments{
\item{x}{GStats object}

\item{formula}{formula}

\item{FUN}{function to apply}

\item{per.site}{normalize window scores by window length (kb)}

\item{...}{additional arguments passed to aggregate}
}
\description{
Aggregate GStats statistics
}
