% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ManticoreRSE.R
\docType{methods}
\name{aggregate,ManticoreRSE-method}
\alias{aggregate,ManticoreRSE-method}
\title{Aggregate ManticoreRSE statistics}
\usage{
\S4method{aggregate}{ManticoreRSE}(x, formula, FUN, which = NULL,
  per.site = FALSE, ...)
}
\arguments{
\item{x}{ManticoreRSE object}

\item{formula}{formula}

\item{FUN}{function to apply}

\item{per.site}{normalize window scores by window length (kb)}

\item{...}{additional arguments passed to aggregate}
}
\description{
Aggregate ManticoreRSE statistics
}
