% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\docType{methods}
\name{aggregate_region_stats,GRanges-method}
\alias{aggregate_region_stats,GRanges-method}
\title{destribeIn aggregate_region_stats Aggregate statistics for GRanges instance}
\usage{
\S4method{aggregate_region_stats}{GRanges}(object, agg.fun = c("sum", "mean"),
  ...)
}
\description{
destribeIn aggregate_region_stats Aggregate statistics for GRanges instance
}
