% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{aggregate_region_stats}
\alias{aggregate_region_stats}
\title{aggregate_region_stats}
\usage{
aggregate_region_stats(object, agg.fun = c("sum", "mean"), ...)
}
\arguments{
\item{object}{object}

\item{agg.fun}{aggregation functions}

\item{...}{additional arguments}
}
\value{
GRanges with augmented values
}
\description{
aggregate_region_stats
}
\details{
Aggregate statistics over regions, where the regions typically are
scaffolds or windows. Aggregation is performed over the list of
components seqnames, key (i.e. statistic), region start, region
end, and region width
}
\author{
Per Unneberg
}
