% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GStats.R
\docType{methods}
\name{asGRanges,GStats-method}
\alias{asGRanges,GStats-method}
\title{Convert GStats to GRanges}
\usage{
\S4method{asGRanges}{GStats}(x, long = TRUE, per.site = FALSE)
}
\arguments{
\item{x}{GStats object}

\item{long}{return long format}

\item{per.site}{calculate per site statistics (divide by width)}
}
\value{
GRanges instance
}
\description{
Convert GStats object to GRanges with values filled by data assay
values
}
\author{
Per Unneberg
}
