% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-GStats.R
\docType{methods}
\name{asGRanges}
\alias{asGRanges}
\alias{asGRanges,GStats-method}
\title{Convert object to GRanges}
\usage{
asGRanges(x, ...)

\S4method{asGRanges}{GStats}(x, long = TRUE, per.site = FALSE)
}
\arguments{
\item{x}{an R object}

\item{...}{additional arguments}

\item{long}{return long format}

\item{per.site}{calculate per site statistics (divide by width) per kb}
}
\value{
GRanges instance
}
\description{
Convert GStats object to GRanges with values filled by data assay
values
}
\author{
Per Unneberg
}
