% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\docType{class}
\name{BcftoolsStats-class}
\alias{BcftoolsStats-class}
\alias{BcftoolsStats}
\title{BcftoolsStats}
\description{
BcftoolsStats object
}
\details{
Representation of bcftools stats output
}
\section{Slots}{

\describe{
\item{\code{ID}}{ID section}

\item{\code{SN}}{summary numbers section}

\item{\code{TSTV}}{transition/transversion section}

\item{\code{SiS}}{singleton stats}

\item{\code{AF}}{stats by non-reference allele frequency}

\item{\code{QUAL}}{stats by quality}

\item{\code{IDD}}{indel distribution}

\item{\code{ST}}{substitution types}

\item{\code{DP}}{depth distribution}
}}

