% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{boxplot.GRanges}
\alias{boxplot.GRanges}
\title{boxplot.GRanges}
\usage{
\method{boxplot}{GRanges}(formula = "value ~ population", data = NULL,
  which = levels(factor(data$key)), per.site = TRUE, ...)
}
\arguments{
\item{per.site}{normalize statistics to per-site values}

\item{...}{arguments to pass to boxplot.pg}
}
