% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{boxplot.pg.diversity}
\alias{boxplot.pg.diversity}
\title{boxplot.pg.diversity}
\usage{
\method{boxplot}{pg.diversity}(data = NULL,
  which = c("nuc.diversity.within", "nuc.F_ST.vs.all", "Pi"), wrap = TRUE,
  ...)
}
