% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{boxplot.pg.F_ST}
\alias{boxplot.pg.F_ST}
\title{boxplot.pg.F_ST}
\usage{
\method{boxplot}{pg.F_ST}(data = NULL, formula = "value ~ seqnames",
  which = c("nucleotide.F_ST", "Nei.G_ST"), wrap = TRUE, ...)
}
