% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{boxplot.pg.neutrality}
\alias{boxplot.pg.neutrality}
\title{boxplot.pg.neutrality}
\usage{
\method{boxplot}{pg.neutrality}(data = NULL, which = c("Tajima.D",
  "Fu.Li.F", "Fu.Li.F"), wrap = TRUE, ...)
}
