% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{boxplot.pg}
\alias{boxplot.pg}
\title{boxplot.pg}
\usage{
\method{boxplot}{pg}(formula = "value ~ population", data = NULL,
  colour = c("black", "gray"), colour.var = NULL, wrap = FALSE,
  wrap.formula = " ~ name", wrap.ncol = 1, plot.type = "box",
  x.lab = "group", y.lab = NULL, main = NULL, compact.facet = TRUE,
  strip.position = "right", scales = "free_y", hide.legend = TRUE,
  text.size = 14, text.x.angle = 45, text.x.hjust = 1, ...)
}
\arguments{
\item{data}{long format from getGenomeStats function}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{plot.type}{plot type, either point or line}

\item{x.lab}{x label}

\item{y.lab}{y label}

\item{main}{plot title}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{hide.legend}{whether or not to hide legend}

\item{text.size}{text size}

\item{text.x.angle}{angle of x tick labels}

\item{text.x.hjust}{horizontal adjustment of x tick labels}

\item{...}{arguments passed on to facet_wrap}

\item{x.var}{variable to map to x aestethic}

\item{y.var}{variable to map to y aestethic}

\item{size}{plot size}

\item{which}{statistic to plot}
}
\value{
ggplot object
}
\description{
boxplot.pg
}
\details{
Make a box/violin plot of data
}
\author{
Per Unneberg
}
