% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gboxplot-GStats.R
\docType{methods}
\name{gboxplot,GStats-method}
\alias{gboxplot,GStats-method}
\title{Make box/violin plot of GStats object}
\usage{
\S4method{gboxplot}{GStats}(data, formula = "value ~ population",
  type = "box", xlim = NULL, ylim = NULL, main = paste(data@statistics,
  "statistics"), xlab = "group", ylab = NULL, colour = brewer.pal(3,
  "Dark2"), colour.var = "seqnames", wrap = FALSE, wrap.formula = "",
  wrap.ncol = 1, compact.facet = TRUE, strip.position = "right",
  scales = "free_y", hide.legend = TRUE, grid = FALSE, text.size = 14,
  text.x.angle = 45, text.x.hjust = 1, which = NULL, per.site = FALSE,
  ...)
}
\arguments{
\item{data}{long format from GStats function}

\item{formula}{a formula, such as y ~ grp}

\item{type}{plot type, either point or line}

\item{main}{plot title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{hide.legend}{whether or not to hide legend}

\item{grid}{include grid lines}

\item{text.size}{text size}

\item{text.x.angle}{angle of x tick labels}

\item{text.x.hjust}{horizontal adjustment of x tick labels}

\item{which}{statistic to plot}

\item{...}{arguments passed on to facet_wrap}

\item{size}{plot size}
}
\value{
ggplot object
}
\description{
Make a box/violin plot of data
}
\author{
Per Unneberg
}
