% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{gboxplot.pg}
\alias{gboxplot.pg}
\alias{gboxplot.pg.detail}
\alias{gboxplot.pg.neutrality}
\alias{gboxplot.pg.fixed}
\alias{gboxplot.pg.shared}
\alias{gboxplot.pg.diversity}
\alias{gboxplot.pg.diversity.between}
\alias{gboxplot.pg.F_ST}
\alias{gboxplot.pg.F_ST.pairwise}
\alias{gboxplot.pg.segregating.sites}
\title{gboxplot.pg}
\usage{
gboxplot.pg(data, x.var = "population", y.var = "value",
  colour = c("black", "gray"), colour.var = NULL, wrap = FALSE,
  wrap.formula = " ~ name", wrap.ncol = 1, plot.type = "box",
  x.lab = "group", y.lab = NULL, main = NULL, compact.facet = TRUE,
  strip.position = "right", scales = "free_y", hide.legend = TRUE,
  text.size = 14, text.x.angle = 45, text.x.hjust = 1, ...)

gboxplot.pg.detail(data, which = c("MDG1", "MDG2", "MDSD"),
  main = "detail statistics", ...)

gboxplot.pg.neutrality(data, which = c("Tajima.D", "Fu.Li.F", "Fu.Li.F"),
  wrap = TRUE, wrap.formula = "~ key", ...)

gboxplot.pg.fixed(data, main = "Fixed sites", ...)

gboxplot.pg.shared(data, main = "Shared sites", ...)

gboxplot.pg.diversity(data, which = c("nuc.diversity.within",
  "nuc.F_ST.vs.all", "Pi"), wrap = TRUE, wrap.formula = "~ key", ...)

gboxplot.pg.diversity.between(data, ...)

gboxplot.pg.F_ST(data, x.var = "name", which = c("nucleotide.F_ST",
  "Nei.G_ST"), wrap = TRUE, wrap.formula = "~ key", ...)

gboxplot.pg.F_ST.pairwise(data, ...)

gboxplot.pg.segregating.sites(data, ...)
}
\arguments{
\item{data}{long format from getGenomeStats function}

\item{x.var}{variable to map to x aestethic}

\item{y.var}{variable to map to y aestethic}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{plot.type}{plot type, either point or line}

\item{x.lab}{x label}

\item{y.lab}{y label}

\item{main}{plot title}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{hide.legend}{whether or not to hide legend}

\item{text.size}{text size}

\item{text.x.angle}{angle of x tick labels}

\item{text.x.hjust}{horizontal adjustment of x tick labels}

\item{...}{arguments passed on to facet_wrap}

\item{which}{statistic to plot}

\item{size}{plot size}
}
\value{
ggplot object
}
\description{
gboxplot.pg
}
\details{
Make a box/violin plot of data
}
\section{Functions}{
\itemize{
\item \code{gboxplot.pg.detail}: Make boxplot of detail statistiscs

\item \code{gboxplot.pg.neutrality}: Make boxplot of netrality statistics

\item \code{gboxplot.pg.fixed}: Make boxplot of fixed sites

\item \code{gboxplot.pg.shared}: Make boxplot of shared sites

\item \code{gboxplot.pg.diversity}: Make boxplot of diversity

\item \code{gboxplot.pg.diversity.between}: Make boxplot of between population diversity

\item \code{gboxplot.pg.F_ST}: Make boxplot of F_ST

\item \code{gboxplot.pg.F_ST.pairwise}: Make boxplot of pairwise F_ST

\item \code{gboxplot.pg.segregating.sites}: Make boxplot of segregating sites
}}

\author{
Per Unneberg
}
