% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/popGenome.R
\docType{class}
\name{GENOMEList-class}
\alias{GENOMEList-class}
\alias{GENOMEList,list-method}
\title{List of GENOME instances}
\usage{
\S4method{GENOMEList}{list}(obj)
}
\description{
Subclass of S4Vectors SimpleList, where each entry is an PopGenome
GENOME instance
}
\section{Methods (by generic)}{
\itemize{
\item \code{GENOMEList}: Convert list to GENOMEList
}}

