% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{GENOMEList}
\alias{GENOMEList}
\title{GENOMEList}
\usage{
GENOMEList(obj, ...)
}
\arguments{
\item{obj}{passed object}

\item{...}{additional arguments}
}
\value{
GENOMEList instance
}
\description{
GENOMEList
}
\details{
Convert objects to GENOMEList
}
\author{
Per Unneberg
}
