% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\docType{methods}
\name{genomewide.stats}
\alias{genomewide.stats}
\alias{genomewide.stats,GENOME-method}
\title{genomewide.stats}
\usage{
genomewide.stats(object, which = c("detail", "neutrality", "F_ST",
  "diversity", "diversity.between", "fixed.shared", "linkage"),
  biallelic.structure = TRUE, ...)

\S4method{genomewide.stats}{GENOME}(object, which = c("detail", "neutrality",
  "F_ST", "diversity", "diversity.between", "fixed.shared", "linkage"),
  biallelic.structure = TRUE, ...)
}
\arguments{
\item{object}{object on which to perform calculations}

\item{which}{statistics to calculate}

\item{biallelic.structure}{calculate biallelic structure}

\item{...}{extra arguments}
}
\value{
Updated object
}
\description{
genomewide.stats
}
\details{
Calculate genome wide statistics
}
\section{Methods (by class)}{
\itemize{
\item \code{GENOME}: Calculate genome wide statistics for a GENOME object
}}

\author{
Per Unneberg
}
