% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get.tree}
\alias{get.tree}
\title{Retrieve trees for testing and examples}
\usage{
get.tree(tree)
}
\arguments{
\item{tree}{Number corresponding to which tree to retrieve (1 or 2)}
}
\value{
A Newick string
}
\description{
Retrieve trees for testing and examples
}
