% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\docType{methods}
\name{getGenomeStats}
\alias{getGenomeStats}
\alias{getGenomeStats,list-method}
\title{getGenomeStats}
\usage{
getGenomeStats(object, stats = c("detail"), use.population.names = FALSE,
  use.region.names = FALSE, long.format = TRUE, quiet = FALSE, ...)

\S4method{getGenomeStats}{list}(object, stats = c("detail"),
  use.population.names = FALSE, use.region.names = FALSE,
  long.format = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{object}{An R object}

\item{stats}{statistic to return}

\item{use.population.names}{use population names of object for plotting labels}

\item{use.region.names}{Use the region names as row names}

\item{long.format}{Return data frame in long output format}

\item{quiet}{suppress messages}

\item{...}{Arguments to pass to data access functions}
}
\value{
A data frame in long or wide format, suitable for plotting with
    ggplot2
}
\description{
Get stats from a PopGenome GENOME object
}
\details{
Retrieve genome stats for a PopGenome GENOME instance. Note that
this is only intended for use with some slots/statistical
operations. The PopGenome function does not work on data that has
been saved in rda files. They should be saved with save.session,
but this does not seem to work.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Retrieve and concatenate data from a list of PopGenome GENOME instances.
}}
\examples{
library(PopGenome)
# Read a vcf and generate two genome objects
vcf_file <- system.file("extdata", "medium.call.vcf.gz")
scaffold1 <- readVCF(vcf_file, 1000, frompos=1, topos=1000000, tid="scaffold1")
scaffold13 <- readVCF(vcf_file, 1000, frompos=1, topos=1000000, tid="scaffold13")
slist <- list(scaffold1=scaffold1, scaffold13=scaffold13)

}
\author{
Per Unneberg
}

