% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot-GStats.R
\docType{methods}
\name{gplot,GStats-method}
\alias{gplot,GStats-method}
\title{Plot GStats objects}
\usage{
\S4method{gplot}{GStats}(data, x = "feature_id", y = "value",
  type = "point", xlim = NULL, ylim = NULL,
  main = paste(data@statistics, "statistics"), xlab = "window",
  ylab = NULL, size = 1, colour = brewer.pal(3, "Dark2"),
  colour.var = "seqnames", wrap = TRUE, wrap.formula = "",
  wrap.ncol = 1, compact.facet = TRUE, strip.position = "right",
  scales = "free_y", hide.legend = TRUE, hide.xaxis = TRUE,
  grid = FALSE, text.size = 14, text.x.angle = 45, text.x.hjust = 1,
  which = NULL, per.site = FALSE, ...)
}
\arguments{
\item{data}{GStats object}

\item{x}{variable to map to x aestethic}

\item{y}{variable to map to y aestethic}

\item{main}{plot title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{size}{plot size}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{hide.legend}{whether or not to hide legend}

\item{hide.xaxis}{hide x axis tick marks and labels}

\item{grid}{include grid lines}

\item{text.size}{text size}

\item{text.x.angle}{x text angle}

\item{text.x.hjust}{x text horizontal justification}

\item{which}{statistic to plot}

\item{...}{extra arguments}

\item{plot.type}{plot type, either point or line}
}
\value{
ggplot
}
\description{
Plot GStats objects
}
\author{
Per Unneberg
}
