% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\name{gplot.bcftools.stats}
\alias{gplot.bcftools.stats}
\title{plot.bcftools.stats}
\usage{
gplot.bcftools.stats(obj, which = c("SN", "TSTV", "SiS", "AF", "QUAL",
  "IDD", "ST", "DP"), ncol = 2, text.size = 12,
  theme.default = theme_bw(), ...)
}
\arguments{
\item{obj}{bcftools.stats object}

\item{which}{which plots to produce (defalt all)}

\item{ncol}{number of columns in grid plot}

\item{text.size}{text size of plots}

\item{theme.default}{default theme (theme_bw)}

\item{...}{parameters passed to generic plot function}
}
\value{
ggplot2 object, or a plot
}
\description{
gplot.bcftools.stats
}
\details{
Plot summary of bcftools stats results
}
\author{
Per Unneberg
}
