% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gplot}
\alias{gplot}
\title{Generic plot function}
\usage{
gplot(data, x = NULL, y = NULL, type = "point", xlim = NULL,
  ylim = NULL, main = NULL, xlab = NULL, ylab = NULL,
  size = integer(0), colour = list(), colour.var = character(0),
  wrap = FALSE, wrap.formula = character(0), wrap.ncol = integer(0),
  compact.facet = TRUE, strip.position = "right", scales = "free_y",
  hide.legend = TRUE, hide.xaxis = TRUE, grid = FALSE,
  text.size = integer(0), text.x.angle = integer(0),
  text.x.hjust = integer(0), which = NULL, ...)
}
\description{
Generic plot function
}
