% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GStats.R
\docType{methods}
\name{GStats,GRanges,ANY-method}
\alias{GStats,GRanges,ANY-method}
\title{Create GStats from GRanges object}
\usage{
\S4method{GStats}{GRanges,ANY}(object, population, statistics,
  application = NA, ...)
}
\description{
Create GStats from GRanges object
}
