% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/popGenome.R
\docType{methods}
\name{GStats}
\alias{GStats}
\alias{GStats,GENOMEList,GRanges_OR_missing-method}
\title{Create GStats object}
\usage{
GStats(object, gr = NULL, statistics = character(0),
  use.population.names = FALSE, use.region.names = FALSE, quiet = FALSE,
  ...)

\S4method{GStats}{GENOMEList,GRanges_OR_missing}(object, gr = NULL,
  statistics = c("detail"), use.population.names = TRUE,
  use.region.names = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{object}{An R object}

\item{gr}{A GRanges object or NULL}

\item{statistics}{statistics to return}

\item{use.population.names}{use population names of object for plotting labels}

\item{use.region.names}{Use the region names as row names}

\item{quiet}{suppress messages}

\item{...}{Arguments to pass to data access functions}
}
\value{
A GStats object
}
\description{
Retrieve genome stats for

Retrieve genome stats for a PopGenome GENOME instance. Note that
this is only intended for use with some slots/statistical
operations. The PopGenome function does not work on data that has
been saved in rda files. They should be saved with save.session,
but this does not seem to work.
}
\examples{
library(PopGenome)
fn <- system.file("extdata", "popgenome.rda", package = "nonmodelr")
load(fn)
gl <- GENOMEList(scaffolds)
sessionInfo()
gs <- GStats(gl)

}
\author{
Per Unneberg
}
