% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{identify_outliers}
\alias{identify_outliers}
\title{identify_outliers}
\usage{
identify_outliers(data, formula, key = NULL, method = "loess",
  level = 0.99, ...)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula to fit}

\item{key}{factor over which to perform fits}

\item{method}{fitting method}

\item{level}{level of confidence interval to use}

\item{...}{additional arguments}
}
\value{
data frame with fitted model values, se, residuals and key
}
\description{
identify_outliers
}
\details{
ientify outliers in a data frame
}
\author{
Per Unneberg
}
