% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-manticore.R
\docType{package}
\name{manticore}
\alias{manticore}
\alias{manticore-package}
\title{manticore: A package for analyses of non-model organism data}
\description{
The manticore package provides functionality for working with non-model organism data
}
\section{manticore functions}{


gplot
}

