% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManticoreDF-class.R
\docType{class}
\name{ManticoreDF-class}
\alias{ManticoreDF-class}
\title{ManticoreDF}
\description{
Derived data frame class that requires that certain metadata
properties be present
}
\author{
Per Unneberg
}
