% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManticoreDF-class.R
\name{ManticoreDF}
\alias{ManticoreDF}
\title{ManticoreDF}
\usage{
ManticoreDF(..., measurement.name = character(),
  application = character())
}
\arguments{
\item{...}{options to pass to DataFrame}

\item{measurement.name}{Measurement name}

\item{application}{Application name}
}
\value{
ManticoreDF object
}
\description{
ManticoreDF initialization function
}
\author{
Per Unneberg
}
