% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManticoreRSE-class.R
\name{ManticoreRSE}
\alias{ManticoreRSE}
\title{ManticoreRSE}
\usage{
ManticoreRSE(..., window.size = integer())
}
\arguments{
\item{...}{options to be passed to RangedSummarizedExperiment}

\item{window.size}{Window size}
}
\value{
ManticoreRSE object
}
\description{
ManticoreRSE initialization function
}
\author{
Per Unneberg
}
