% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\name{plot.bcftools.stats}
\alias{plot.bcftools.stats}
\title{plot.bcftools.stats}
\usage{
plot.bcftools.stats(obj, which = c("SN", "TSTV", "SiS", "AF", "QUAL", "IDD",
  "ST", "DP"), ncol = 2, ...)
}
\arguments{
\item{obj}{bcftools.stats object}

\item{which}{which plots to produce (defalt all)}

\item{ncol}{number of columns in grid plot}

\item{...}{parameters passed to generic plot function}
}
\value{
ggplot2 object, or a plot
}
\description{
plot.bcftools.stats
}
\details{
Plot summary of bcftools stats results
}
\author{
Per Unneberg
}
