% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{plot.GRanges}
\alias{plot.GRanges}
\title{plot.GRanges}
\usage{
\method{plot}{GRanges}(data, which = levels(factor(data$key)),
  per.site = TRUE, y = "value", ...)
}
\arguments{
\item{per.site}{normalize statistics to per-site values}
}
