% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{plot.pg.diversity.between}
\alias{plot.pg.diversity.between}
\title{plot.pg.diversity.between}
\usage{
\method{plot}{pg.diversity.between}(data, wrap.formula = "~ population",
  colour = "name", ...)
}
