% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{plot.pg.summary}
\alias{plot.pg.summary}
\title{plot.pg.summary}
\usage{
\method{plot}{pg.summary}(data, x = "seqnames", y = "value",
  wrap.formula = "~ key", which = c("n.sites", "trans.transv.ratio"),
  x.lab = "scaffold", hide.xaxis = FALSE, ...)
}
