% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenome.R
\docType{methods}
\name{plot_region_stats}
\alias{plot_region_stats}
\alias{plot_region_stats,GRanges-method}
\title{plot_region_stats}
\usage{
plot_region_stats(object, x.var = "width", wrap = TRUE,
  agg.fun = c("sum", "mean"), text.size = 12, which = NULL,
  label.outlier = FALSE, label.format = "short", n.se = 5,
  label.hjust = 1, label.vjust = 1, level = 0.99, method = "loess",
  scales = "free_y", ...)

\S4method{plot_region_stats}{GRanges}(object, x.var = "width",
  wrap = TRUE, agg.fun = c("sum", "mean"), text.size = 12,
  which = NULL, label.outlier = FALSE, label.format = "short",
  n.se = 5, label.hjust = 1, label.vjust = 1, level = 0.99,
  method = "loess", scales = "free_y", ...)
}
\arguments{
\item{object}{object}

\item{x.var}{a name identifying a variable to plot values against}

\item{wrap}{wrap plot}

\item{agg.fun}{aggregation functions}

\item{text.size}{text size}

\item{which}{which plots to include. Choose from population levels and aggregation function}

\item{label.outlier}{label outlier points}

\item{label.format}{use short (default) or long format for window labels}

\item{n.se}{number of standard errors to use as cutoff for identifying outliers, based on confidence interval level}

\item{label.hjust}{horizontal adjustment for labels}

\item{label.vjust}{vertical adjustment for labels}

\item{level}{level of confidence interval to use}

\item{method}{method to use for fitting}

\item{scales}{scales for facet plots}

\item{...}{additional arguments to pass to geom_point()}
}
\value{
ggplot
}
\description{
plot_region_stats
}
\details{
Plot region stats
}
\section{Methods (by class)}{
\itemize{
\item \code{GRanges}: Plot statistics versus region for GRanges instance
}}

\author{
Per Unneberg
}
