% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\name{read.bcftools.stats}
\alias{read.bcftools.stats}
\title{read.bcftools.stats}
\usage{
read.bcftools.stats(filename, label = NULL)
}
\arguments{
\item{filename}{File to read}

\item{label}{Label to assign object (e.g. sample)}
}
\value{
bcftools.stats object
}
\description{
read.bcftools.stats
}
\details{
Read output from bcftools stats
}
\author{
Per Unneberg
}

