% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\name{readBcftoolsStats}
\alias{readBcftoolsStats}
\title{readBcftoolsStats}
\usage{
readBcftoolsStats(filename, label = NULL)
}
\arguments{
\item{filename}{File to read}

\item{label}{Label to assign object (e.g. sample)}
}
\value{
BcftoolsStats object
}
\description{
readBcftoolsStats
}
\details{
Read output from bcftools stats
}
\author{
Per Unneberg
}
