% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popoolation.R
\name{readVarianceSliding}
\alias{readVarianceSliding}
\title{readVarianceSliding}
\usage{
readVarianceSliding(filename, measure = "pi")
}
\arguments{
\item{filename}{filename to parse}

\item{measure}{measure}
}
\value{

}
\description{
Read VarianceSliding results from popoolation
}
\author{
Per Unneberg
}
