% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_test.R
\name{rtc.sensitivity.test}
\alias{rtc.sensitivity.test}
\title{Perform random tree congruence test}
\usage{
rtc.sensitivity.test(
  reference.tree,
  comparison.tree,
  null.congruence.model,
  metric,
  spr.proportions = seq(0, 0.5, by = 0.05),
  iterations = 100
)
}
\arguments{
\item{reference.tree}{Path to reference tree file, or string in Newick format}

\item{comparison.tree}{Path to comparison tree file, or string in Newick format}

\item{null.congruence.model}{Null model generated by rtc.test function}

\item{metric}{Congruence metric specified for null model generation}

\item{spr.proportions}{Vector of proportions of tree size to subtree prune and regraft: default = seq(0, 0.5, by=0.05)}

\item{iterations}{Number of times to perform SPR on input trees to generate congruence distribution: default = 100}
}
\value{
Function returns a dataframe containing the integrated P(Null >= Observed)
(integrated.p column) as a function of SPR proportion (prop.dive column).
}
\description{
Run a sensitivity test on random tree congruence test
results. This function tests how sensitive RTC test results
are to variation in congruence estimates (uncertainty). To
accomplish this, a distribution of congruence values are
generated by iteratively using subtree pruning (SPR) and regrafting to mutate
the original input trees. The congruence distribution is then compared
against the null model, and an integrated P(Null >= Observed) values
is calculated across the congruence distribution.
The extent of SPR is increased to introduce greater variation.
}
