% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\name{summary.bcftools.stats}
\alias{summary.bcftools.stats}
\title{summary.bcftools.stats}
\usage{
\method{summary}{bcftools.stats}(obj)
}
\arguments{
\item{obj}{bcftools.stats object}
}
\value{
data frame summary
}
\description{
summary.bcftools.stats
}
\details{
Print summary for bcftools stats
}
\author{
Per Unneberg
}

