% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{summary.GRanges}
\alias{summary.GRanges}
\title{summary.GRanges}
\usage{
\method{summary}{GRanges}(gr, per.site = TRUE, fun = "mean", ...)
}
\arguments{
\item{gr}{GRanges object}

\item{per.site}{calculate per site values}

\item{...}{additional arguments}
}
\value{
a data frame
}
\description{
summary.GRanges
}
\details{
Summarize the statistics of a GRanges object
}
\author{
Per Unneberg
}
