% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ManticoreRSE.R
\name{summary.ManticoreRSE}
\alias{summary.ManticoreRSE}
\title{summary.ManticoreRSE}
\usage{
\method{summary}{ManticoreRSE}(obj, which = NULL, fun = "mean",
  per.site = FALSE, per.region = FALSE, variable = "score",
  sites = "sites", ...)
}
\arguments{
\item{obj}{ManticoreRSE object}

\item{which}{which assay to summarize}

\item{fun}{summary function}

\item{per.region}{normalize score by window size}

\item{...}{additional arguments}

\item{formula}{formula to calculate}
}
\value{
a data frame
}
\description{
Summarize the statistics of a ManticoreRSE object
}
\author{
Per Unneberg
}
