% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWindows-class.R
\docType{class}
\name{SWindows-class}
\alias{SWindows-class}
\alias{SWindows}
\title{SWindows class}
\usage{
SWindows(..., coverage = numeric(), sites = integer(),
  segregating.sites = integer())

SWindows(..., coverage = numeric(), sites = integer(),
  segregating.sites = integer())
}
\arguments{
\item{...}{arguments to pass on to parent class Windows}

\item{coverage}{coverage per window}

\item{sites}{number of good sites per window}

\item{segregating.sites}{number of segregating sites per window}

\item{...}{arguments to pass on to parent class Windows}

\item{coverage}{coverage per window}

\item{sites}{number of good sites per window}

\item{segregating.sites}{number of segregating sites per window}
}
\value{
SWindows class

SWindows class
}
\description{
Class for windows carrying additional parallel slots with
information about number of good sites, window coverage and
segregating sites

Create an SWindows object

Class for windows carrying additional parallel slots with
information about number of good sites, window coverage and
segregating sites

Create an SWindows object
}
\author{
Per Unneberg

Per Unneberg

Per Unneberg

Per Unneberg
}
