% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popoolation.R
\name{VarianceSlidingAssays}
\alias{VarianceSlidingAssays}
\title{VarianceSlidingAssays}
\usage{
VarianceSlidingAssays(input.df, colData = NULL, window.size = NULL,
  class = NA, ...)
}
\arguments{
\item{input.df}{Input data fram that must have columns filename,
sample, and measure}

\item{colData}{column data that describes the samples}

\item{window.size}{window size; if NULL, entire chromosomes are
implied}

\item{class}{arbitrary classification of windows, e.g. to
autosomes, sex chromosomes etc.}

\item{...}{extra parameters passed to readVarianceSliding}
}
\value{
ManticoreRSE Assays consist of the requested statistics
}
\description{
Read multiple VarianceSliding results from popoolation
}
\author{
Per Unneberg
}
