% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Windows-class.R
\docType{class}
\name{Windows-class}
\alias{Windows-class}
\alias{Windows}
\title{Windows representation}
\usage{
Windows(..., window.size = integer())

Windows(..., window.size = integer())
}
\arguments{
\item{...}{parameters to be passed on to parent class GRanges}

\item{window.size}{window size of analysis}

\item{...}{parameters to be passed on to parent class GRanges}

\item{window.size}{window size of analysis}
}
\value{
Windows object

Windows object
}
\description{
Windows object to represent windowed analyses. Contains a slot for
window size. Note that edge effects are allowed such that the last
windows may be smaller than the actual window size.

Create Windows class

Windows object to represent windowed analyses. Contains a slot for
window size. Note that edge effects are allowed such that the last
windows may be smaller than the actual window size.

Create Windows class
}
\author{
Per Unneberg

Per Unneberg

Per Unneberg

Per Unneberg
}
