% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gboxplot-GStats.R
\docType{methods}
\name{gboxplot}
\alias{gboxplot}
\alias{gboxplot,GStats-method}
\title{Generic boxplot function}
\usage{
gboxplot(data, formula = character(0), type = character(0),
  xlim = NULL, ylim = NULL, main = NULL, xlab = NULL,
  ylab = NULL, colour = list(), colour.var = character(0),
  wrap = FALSE, wrap.formula = chacarter(0), wrap.ncol = integer(0),
  compact.facet = TRUE, strip.position = "right", scales = "free_y",
  hide.legend = TRUE, grid = FALSE, text.size = integer(0),
  text.x.angle = integer(0), text.x.hjust = integer(0), which = NULL,
  ...)

\S4method{gboxplot}{GStats}(data, formula = "value ~ population",
  type = "box", xlim = NULL, ylim = NULL,
  main = paste(data@statistics, "statistics"), xlab = "group",
  ylab = NULL, colour = brewer.pal(3, "Dark2"),
  colour.var = "seqnames", wrap = FALSE, wrap.formula = "",
  wrap.ncol = 1, compact.facet = TRUE, strip.position = "right",
  scales = "free_y", hide.legend = TRUE, grid = FALSE,
  text.size = 14, text.x.angle = 45, text.x.hjust = 1,
  which = NULL, per.site = FALSE, ...)
}
\arguments{
\item{data}{long format from GStats function}

\item{formula}{a formula, such as y ~ grp}

\item{type}{plot type, either point or line}

\item{xlim}{x limit}

\item{ylim}{y limit}

\item{main}{plot title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{hide.legend}{whether or not to hide legend}

\item{grid}{include grid lines}

\item{text.size}{text size}

\item{text.x.angle}{angle of x tick labels}

\item{text.x.hjust}{horizontal adjustment of x tick labels}

\item{which}{statistic to plot}

\item{...}{arguments passed on to facet_wrap}

\item{per.site}{normalize statistic by window length}
}
\value{
ggplot object
}
\description{
Make a box/violin plot of data
}
\author{
Per Unneberg
}
