% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/popGenome.R
\docType{methods}
\name{GENOMEList}
\alias{GENOMEList}
\alias{GENOMEList,list-method}
\title{GENOMEList}
\usage{
GENOMEList(obj, ...)

\S4method{GENOMEList}{list}(obj)
}
\arguments{
\item{obj}{an R object}

\item{...}{additional arguments}
}
\description{
GENOMEList

Convert list to GENOMEList
}
