% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenome.R
\name{genomewide.stats}
\alias{genomewide.stats}
\title{genomewide.stats}
\usage{
genomewide.stats(object, which = character(0), biallelic.structure = TRUE,
  pi = TRUE, ...)
}
\description{
genomewide.stats
}
