% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenome.R
\docType{methods}
\name{genomewide.stats}
\alias{genomewide.stats}
\alias{genomewide.stats,GENOME-method}
\title{genomewide.stats}
\usage{
genomewide.stats(object, which = character(0), biallelic.structure = TRUE,
  pi = TRUE, ...)

\S4method{genomewide.stats}{GENOME}(object, which = c("detail", "neutrality",
  "F_ST", "diversity", "diversity.between", "fixed.shared"),
  biallelic.structure = TRUE, pi = TRUE, ...)
}
\arguments{
\item{object}{GENOME object on which to perform calculations}

\item{which}{statistics to calculate}

\item{biallelic.structure}{calculate biallelic structure}

\item{pi}{Nei's calculation of pi}

\item{...}{extra arguments}
}
\value{
Updated object
}
\description{
genomewide.stats

Calculate genome wide statistics
}
\author{
Per Unneberg
}
