% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{genomewide.stats}
\alias{genomewide.stats}
\title{genomewide.stats}
\usage{
genomewide.stats(object, which = c("detail", "neutrality", "F_ST",
  "diversity", "diversity.between", "fixed.shared", "linkage"),
  biallelic.structure = TRUE, ...)
}
\arguments{
\item{object}{GENOME object on which to perform calculations}

\item{which}{statistics to calculate}

\item{biallelic.structure}{calculate biallelic structure}

\item{...}{extra arguments}
}
\value{
Updated object
}
\description{
genomewide.stats
}
\details{
Calculate genome wide statistics
}
\author{
Per Unneberg
}

