% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gplot-GStats.R
\docType{methods}
\name{gplot}
\alias{gplot}
\alias{gplot,GStats-method}
\title{Generic plot function}
\usage{
gplot(data, x = NULL, y = NULL, type = "point", xlim = NULL,
  ylim = NULL, main = NULL, xlab = NULL, ylab = NULL,
  size = integer(0), colour = list(), colour.var = character(0),
  wrap = FALSE, wrap.formula = character(0), pos = character(0),
  wrap.ncol = integer(0), compact.facet = TRUE,
  strip.position = "right", scales = "free_y", hide.legend = TRUE,
  hide.xaxis = TRUE, grid = FALSE, text.size = integer(0),
  text.x.angle = integer(0), text.x.hjust = integer(0), which = NULL,
  per.site = FALSE, zscore = FALSE, ...)

\S4method{gplot}{GStats}(data, x = "feature_id", y = "value",
  type = "point", xlim = NULL, ylim = NULL,
  main = paste(data@statistics, "statistics"), xlab = "bp",
  ylab = NULL, size = 1, colour = brewer.pal(3, "Dark2"),
  colour.var = "seqnames", wrap = TRUE, wrap.formula = "",
  pos = "coordinate", wrap.ncol = 1, compact.facet = TRUE,
  strip.position = "right", scales = "free_y", hide.legend = TRUE,
  hide.xaxis = TRUE, grid = FALSE, text.size = 14,
  text.x.angle = 45, text.x.hjust = 1, which = NULL,
  per.site = FALSE, zscore = FALSE, ...)
}
\arguments{
\item{data}{GStats object}

\item{x}{variable to map to x aestethic}

\item{y}{variable to map to y aestethic}

\item{type}{plot type}

\item{xlim}{x limit}

\item{ylim}{y limit}

\item{main}{plot title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{size}{plot size}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{hide.legend}{whether or not to hide legend}

\item{hide.xaxis}{hide x axis tick marks and labels}

\item{grid}{include grid lines}

\item{text.size}{text size}

\item{text.x.angle}{x text angle}

\item{text.x.hjust}{x text horizontal justification}

\item{which}{statistic to plot}

\item{per.site}{plot statistics normalized by window length}

\item{zscore}{Plot Z-scores}

\item{...}{extra arguments}
}
\value{
ggplot
}
\description{
Plot GStats objects
}
\section{Methods (by class)}{
\itemize{
\item \code{GStats}: 
}}

\author{
Per Unneberg
}
