% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{plot.pg}
\alias{plot.pg}
\title{plot.pg}
\usage{
\method{plot}{pg}(data, x = "pos", y = "value", colour = c("black",
  "gray"), colour.var = "name", wrap = TRUE,
  wrap.formula = "key ~ population", wrap.ncol = 1, plot.type = "point",
  x.lab = "window", y.lab = NULL, main = NULL, colour.scale = NULL,
  compact.facet = TRUE, strip.position = "right", scales = "free_y",
  size = 1, hide.legend = TRUE, text.size = 14, ...)
}
\arguments{
\item{data}{long format from getGenomeStats function}

\item{x}{variable to map to x aestethic}

\item{y}{variable to map to y aestethic}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{plot.type}{plot type, either point or line}

\item{x.lab}{x label}

\item{y.lab}{y label}

\item{main}{plot title}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{size}{plot size}

\item{hide.legend}{whether or not to hide legend}

\item{text.size}{text size}

\item{...}{extra arguments}

\item{name.colours}{colours to use}
}
\value{
ggplot
}
\description{
plot.pg
}
\details{
Generic plotting function for PopGenome results
}
\author{
Per Unneberg
}

