% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\name{plot.pg}
\alias{plot.pg}
\title{plot.pg}
\usage{
\method{plot}{pg}(data, x = "ranges", y = "value", colour = brewer.pal(3,
  "Dark2"), colour.var = "seqnames", wrap = TRUE,
  wrap.formula = "key ~ population", wrap.ncol = 1, plot.type = "point",
  x.lab = "window", y.lab = NULL, main = NULL, compact.facet = TRUE,
  strip.position = "right", scales = "free_y", size = 1,
  hide.legend = TRUE, hide.xaxis = TRUE, grid = FALSE, text.size = 14,
  text.x.angle = 45, text.x.hjust = 1, ...)
}
\arguments{
\item{data}{long format from getGenomeStats function}

\item{x}{variable to map to x aestethic}

\item{y}{variable to map to y aestethic}

\item{colour}{colours to use}

\item{colour.var}{variable to map to colour aestethic}

\item{wrap}{wrap plots}

\item{wrap.formula}{wrap formula}

\item{wrap.ncol}{number of columns in facet wrap}

\item{plot.type}{plot type, either point or line}

\item{x.lab}{x label}

\item{y.lab}{y label}

\item{main}{plot title}

\item{compact.facet}{compact facet representation}

\item{strip.position}{strip position}

\item{scales}{scales}

\item{size}{plot size}

\item{hide.legend}{whether or not to hide legend}

\item{hide.xaxis}{hide x axis tick marks and labels}

\item{grid}{include grid lines}

\item{text.size}{text size}

\item{text.x.angle}{x text angle}

\item{text.x.hjust}{x text horizontal justification}

\item{...}{extra arguments}

\item{which}{statistic to plot}
}
\value{
ggplot
}
\description{
plot.pg
}
\details{
Generic plotting function for PopGenome results
}
\author{
Per Unneberg
}
