% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\docType{methods}
\name{plot_seqlengths_stats}
\alias{plot_seqlengths_stats}
\alias{plot_seqlengths_stats,GRanges-method}
\title{plot_seqlengths_stats}
\usage{
plot_seqlengths_stats(object, wrap = TRUE, agg.fun = c("sum", "mean"),
  theme.default = theme_bw(), text.size = 12, ...)

\S4method{plot_seqlengths_stats}{GRanges}(object, wrap = TRUE,
  agg.fun = c("sum", "mean"), theme.default = theme_bw(), text.size = 12,
  ...)
}
\arguments{
\item{object}{GRanges object}
}
\value{
ggplot
}
\description{
plot_seqlengths_stats
}
\details{
Plot seqlengths stats
}
\section{Methods (by class)}{
\itemize{
\item \code{GRanges}: Plot statistics versus seqlengths for GRanges instance
}}

\author{
Per Unneberg
}
