% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popGenomeUtils.R
\docType{methods}
\name{plot_seqlengths_stats}
\alias{plot_seqlengths_stats}
\alias{plot_seqlengths_stats,GRanges-method}
\title{plot_seqlengths_stats}
\usage{
plot_seqlengths_stats(object, wrap = TRUE, agg.fun = c("sum", "mean"),
  text.size = 12, which = NULL, label.outlier = FALSE, n.se = 5,
  label.hjust = 1, label.vjust = 1, ...)

\S4method{plot_seqlengths_stats}{GRanges}(object, wrap = TRUE,
  agg.fun = c("sum", "mean"), text.size = 12, which = NULL,
  label.outlier = FALSE, n.se = 5, label.hjust = 1, label.vjust = 1,
  ...)
}
\arguments{
\item{object}{GRanges object}

\item{wrap}{wrap plot}

\item{agg.fun}{aggregation functions}

\item{text.size}{text size}

\item{which}{which plots to include. Choose from population levels and aggregation function}

\item{label.outlier}{label outlier points}

\item{n.se}{number of standard errors to use as cutoff for identifying outliers}

\item{label.hjust}{horizontal adjustment for labels}

\item{label.vjust}{vertical adjustment for labels}

\item{...}{additional arguments to pass to geom_point()}
}
\value{
ggplot
}
\description{
plot_seqlengths_stats
}
\details{
Plot seqlengths stats
}
\section{Methods (by class)}{
\itemize{
\item \code{GRanges}: Plot statistics versus seqlengths for GRanges instance
}}

\author{
Per Unneberg
}
