% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popoolation.R
\name{readVarianceSliding}
\alias{readVarianceSliding}
\title{readVarianceSliding}
\usage{
readVarianceSliding(filename, measure = "pi", sample, seqinfo = NULL,
  window.size = integer(), ...)
}
\arguments{
\item{filename}{filename to parse}

\item{measure}{measure}

\item{sample}{sample name}

\item{seqinfo}{Seqinfo object}

\item{window.size}{window size}
}
\value{
WindowedSummarizedExperiment object of one assay
}
\description{
Read VarianceSliding results from popoolation. This function reads
data, adds start and end columns and converts data to a
WindowedSummarizedExperiment object.
}
\examples{
fn <- system.file("extdata", "popoolation", "dmel.A.D.txt.gz", package = "manticore")
readVarianceSliding(fn, measure = "D", sample = "A")

}
\author{
Per Unneberg
}
