% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popoolation.R
\name{readVarianceSliding}
\alias{readVarianceSliding}
\title{readVarianceSliding}
\usage{
readVarianceSliding(filename, measure = "pi", window.size = NULL,
  sample)
}
\arguments{
\item{filename}{filename to parse}

\item{measure}{measure}

\item{window.size}{window size}
}
\value{

}
\description{
Read VarianceSliding results from popoolation. This function reads
data, adds start and end columns and converts data to a
ManticoreRSE object.
}
\author{
Per Unneberg
}
