% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_tree_congruence_test.R
\name{rtc.test}
\alias{rtc.test}
\title{Random Tree Congruence Test}
\usage{
rtc.test(
  reference.tree,
  comparison.tree,
  congruence.metric,
  iterations,
  verbose = FALSE
)
}
\arguments{
\item{reference.tree}{Path to reference tree file, or string in Newick format}

\item{comparison.tree}{Path to comparison tree file, or string in Newick format}

\item{congruence.metric}{Metric used to evaluate congruence. Options include:
\itemize{
\item MCI: Mutual Clustering Information
\item SPI: Shared Phylogenetic Information
\item NS: Nye Similarity
\item JRF: Jaccard-Robinson-Foulds
\item MSD: Matching Split Distance
\item MSID: Matching Split Information Distance
\item RF: Robinson-Foulds
\item ICRF: Information-Corrected Robinson Foulds
}

Note: add '+n' to metric abbreviation to perform normalization. For example,
for normalized JRF, use 'JRF+n'}

\item{iterations}{The number of randomly simulated trees used to construct null distribution}

\item{verbose}{Display run updates (TRUE of FALSE)}
}
\value{
Function returns the random tree congruence test results. This is a list object that contains:

\if{html}{\out{<b>}}observed.congruence\if{html}{\out{</b>}}: The observed congruence between the reference and comparison trees (based on specified congruence metric)

\if{html}{\out{<b>}}p\if{html}{\out{</b>}}: The proportion of values in the null distribution that are greater than or equal to the observed congruence

\if{html}{\out{<b>}}null.congruence.model\if{html}{\out{</b>}}: A vector containing the null congruence values
}
\description{
Random Tree Congruence Test
}
