% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GStats.R
\name{summary.GStats}
\alias{summary.GStats}
\title{summary.GStats}
\usage{
\method{summary}{GStats}(gs, per.site = TRUE, fun = "mean",
  per.region = FALSE, ...)
}
\arguments{
\item{gs}{GStats object}

\item{per.site}{calculate per site values}

\item{fun}{summary function}

\item{per.region}{normalize score by window size}

\item{...}{additional arguments}
}
\value{
a data frame
}
\description{
Summarize the statistics of a GStats object by regions or
populations
}
\author{
Per Unneberg
}
