% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popoolation.R
\name{VarianceSlidingAssays}
\alias{VarianceSlidingAssays}
\title{VarianceSlidingAssays}
\usage{
VarianceSlidingAssays(input.df, colData = NULL, window.size = NULL,
  class = NA, ...)
}
\arguments{
\item{input.df}{Input data fram that must have columns filename,
sample, and measure}

\item{colData}{column data that describes the samples}

\item{window.size}{window size; if NULL, entire chromosomes are
implied}

\item{class}{arbitrary classification of windows, e.g. to
autosomes, sex chromosomes etc.}

\item{...}{extra parameters passed to readVarianceSliding}
}
\value{
WindowedSummarizedExperiment Assays consist of the requested statistics
}
\description{
Read multiple VarianceSliding results from popoolation
}
\examples{
dmel.df <- data.frame(filename = list.files(system.file("extdata", "popoolation", package = "manticore"), full.names = TRUE),
                      sample = c(rep("A", 3), rep("B", 3)), measure = rep(c("D", "pi", "theta"), 2))
mrse <- VarianceSlidingAssays(dmel.df)


}
\author{
Per Unneberg
}
